\docType{package}
\name{brainGL-package}
\alias{brainGL}
\alias{brainGL-package}
\title{Functions to render 3D brain images in html}
\description{
  Functions to render 3D brain images in html
}
\author{
  \email{muschellij2@gmail.com}
}
\seealso{
  \code{\link{misc3d}}, \code{\link{rgl}}
}
\keyword{brainGL}
\keyword{package,}

