% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeScene.R
\name{makeScene}
\alias{makeScene}
\title{Make Leveled Scene}
\usage{
makeScene(data, cutoffs, alpha, cols)
}
\arguments{
\item{data}{- 3D array of values (can be \link{nifti-class})}

\item{cutoffs}{- series of levels to be created}

\item{alpha}{- alpha levels for each contour}

\item{cols}{- colors for each contour}
}
\value{
scene with multiple objects - can be passed to \link{write4D}
}
\description{
Make scene returns a list of levels - but makes them mutually distinct.  
So if cutoff 0.1, 0.2, then 0.1<= x < 0.2 is an roi, not > 0.1 and > 0.2.
Different than \code{\link{contour3d}} as these are mutually exclusive levels.
}

