\docType{data}
\name{MNI_Template}
\alias{MNI_Template}
\title{Information
Brain Template from Copyright (C) 1993–2009 Louis Collins, McConnell Brain Imaging Centre,
Montreal Neurological Institute, McGill University
6th generation non-linear symmetric brain}
\description{
  Information Brain Template from Copyright (C) 1993–2009
  Louis Collins, McConnell Brain Imaging Centre, Montreal
  Neurological Institute, McGill University 6th generation
  non-linear symmetric brain
}
\references{
  \url{http://www.bic.mni.mcgill.ca/ServicesAtlases/ICBM152NLin6}
}
\keyword{data}
\keyword{datasets}

